Remove (and rename) package org.apache.spark.sql.delta.shims
