+++ spark/src/main/scala-spark-3.5/shims/ColumnDefinitionShim.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.delta.shims
+package org.apache.spark.sql.catalyst.plans.logical
 
 import org.apache.spark.sql.catalyst.parser.ParserInterface
 import org.apache.spark.sql.types.{StructField, StructType}
