+++ spark/src/main/scala-spark-master/shims/ColumnDefinitionShim.scala
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.delta.shims
+package org.apache.spark.sql.catalyst.plans.logical
 
 import org.apache.spark.sql.catalyst.parser.ParserInterface
-import org.apache.spark.sql.catalyst.plans.logical.ColumnDefinition
 import org.apache.spark.sql.types.StructField
 
 object ColumnDefinitionShim {
